backend/requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
openai>=1.78.0
python-dotenv>=1.0.0
pydantic>=2.5.2
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
httpx>=0.26.0
python-jose>=3.3.0
passlib>=1.7.4